refactor: input 리팩토링, option 상수로 저장 #15

+++ src/containers/Comment/CommentList/CommentList.tsx
@@ -1,16 +1,25 @@
 "use client";
 
-import React, { useState, useEffect, MouseEvent } from "react";
+import React, { useState, useEffect } from "react";
 import styles from "./CommentList.module.scss";
 import { CommentListProps, Comments } from "@/types/post";
 import { getDateForm } from "@/utils/getDateForm";
 import { checkBlogAdmin } from "@/utils/sessionCheck/checkBlogAdmin";
+import { COMMENT_FORM_LENGTH } from "@/constants/commentConstants";
+import { CustomInput, CustomTextarea } from "@/components/inputs/CustomInputs/CustomInputs";
 
 const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
-  const [commentList, setCommentList] = useState<Comments[]>([]);
+  const { MIN_NICKNAME, MIN_PASSWORD, MIN_COMMENT, MAX_NICKNAME, MAX_PASSWORD, MAX_COMMENT } =
+    COMMENT_FORM_LENGTH;
 
-  const [editComment, setEditComment] = useState<string>("");
-  const [editingCommentId, setEditingCommentId] = useState<string>("");
+  const [commentList, setCommentList] = useState<Comments[]>([]); // API 요청하여 조회할 댓글 목록
+
+  const [editComment, setEditComment] = useState<string>(""); // 수정 input
+  const [editingCommentId, setEditingCommentId] = useState<string>(""); // 수정중인 댓글 ObjectId
+
+  const [checkingGuestPassword, setCheckingGuestPassword] = useState<boolean>(false); // 게스트 댓글 비밀번호 input이 나타날지 말지 여부
+  const [deletePassword, setDeletePassword] = useState<string>(""); // 게스트 댓글 비밀번호 input
+  const [deletingCommentId, setDeletingCommentId] = useState<string>(""); // 수정중인 댓글 ObjectId
 
   // 게시물의 댓글을 조회하여 state에 저장합니다.
   useEffect(() => {
@@ -31,23 +40,21 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
 
   // 수정 버튼 클릭 이벤트
   const handleClickEditButton = async (_id: string, originComment: string) => {
-    // 수정할 코멘트 정보
-    setEditingCommentId(_id);
-    setEditComment(originComment);
+    setEditingCommentId(_id); // 수정할 코멘트 id
+    setEditComment(originComment); // 코멘트 불러오기
   };
 
   // 수정 취소 버튼 클릭 이벤트
   const handleClickCancelEditButton = () => {
-    // 수정할 코멘트 정보 초기화
-    setEditingCommentId("");
-    setEditComment("");
+    setEditingCommentId(""); // 수정할 코멘트 id 초기화
+    setEditComment(""); // 코멘트 초기화
   };
 
   // 수정 확인 버튼 클릭 이벤트
-  const handleClickConfirmEditButton = async (commentId: string) => {
+  const handleClickConfirmEditButton = async (_id: string) => {
     try {
       // PATCH 요청을 보냅니다.
-      const response = await fetch(`/api/posts/${postId}/comments?_id=${commentId}`, {
+      const response = await fetch(`/api/posts/${postId}/comments?_id=${_id}`, {
         method: "PATCH",
         headers: {
           "Content-Type": "application/json",
@@ -66,26 +73,28 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
       // 수정한 댓글을 반영한 결과를 state에 저장합니다.
       const copiedComments: Comments[] = [...commentList];
       const editedComment: Comments | undefined = copiedComments.find(
-        (comment) => String(comment._id) === commentId
+        (comment) => String(comment._id) === _id
       );
       if (editedComment) {
         editedComment.comment = editComment;
       }
       setCommentList(copiedComments);
 
-      // 수정 코멘트 정보 초기화
-      setEditingCommentId("");
-      setEditComment("");
+      setEditingCommentId(""); // 수정할 코멘트 id 초기화
+      setEditComment(""); // 코멘트 초기화
     } catch (err) {
       console.error(err);
     }
   };
 
   // 삭제 버튼 클릭 이벤트
-  const handleClickDeleteButton = async (commentId: string) => {
+  const handleClickDeleteButton = async (_id: string) => {
+    // 댓글 삭제 확인
+    if (!window.confirm("정말로 댓글을 삭제하시겠습니까?")) return;
+
     try {
       // GET 요청을 보냅니다.
-      const response = await fetch(`/api/posts/${postId}/comments?_id=${commentId}`, {
+      const response = await fetch(`/api/posts/${postId}/comments?_id=${_id}`, {
         method: "DELETE",
       });
 
@@ -97,7 +106,7 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
 
       // 삭제한 댓글을 제외한 결과를 state에 저장합니다.
       const afterDeleteComments: Comments[] = commentList.filter(
-        (comment: Comments) => String(comment._id) !== commentId
+        (comment: Comments) => String(comment._id) !== _id
       );
       setCommentList(afterDeleteComments);
     } catch (err) {
@@ -105,18 +114,18 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
     }
   };
 
-  const [checkingGuestPassword, setCheckingGuestPassword] = useState<boolean>(false);
-  const [deletePassword, setDeletePassword] = useState<string>("");
-
   // 게스트 댓글 삭제 버튼 클릭
-  const handleClickGuestDeleteButton = () => {
+  const handleClickGuestDeleteButton = (_id: string) => {
     setCheckingGuestPassword(true);
-    setEditingCommentId("");
+    setDeletingCommentId(_id);
     setEditComment("");
   };
 
   // 게스트 비밀번호 입력 후 삭제확인 버튼 클릭
-  const handleClickGuestPasswordConfirm = async (_id: string) => {
+  const handleClickConfirmGuestPassword = async (_id: string) => {
+    // 댓글 삭제 확인
+    if (!window.confirm("정말로 댓글을 삭제하시겠습니까?")) return;
+
     try {
       // POST 요청을 보냅니다.
       const response = await fetch(`/api/posts/${postId}/comments/guest?_id=${_id}`, {
@@ -155,6 +164,18 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
     setDeletePassword("");
   };
 
+  const deletePasswordInputProps = {
+    value: deletePassword,
+    maxLength: MAX_PASSWORD,
+    dispatch: setDeletePassword,
+  };
+
+  const editCommentInputProps = {
+    value: editComment,
+    maxLength: MAX_COMMENT,
+    dispatch: setEditComment,
+  };
+
   return (
     <ul className={styles.commentList}>
       {commentList.map((commentItem: Comments) => {
@@ -166,7 +187,8 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
         const isSameCommenter: boolean = isLoggedIn && userEmail === author; // 동일한 댓글 작성자
         const isBlogAdmin: boolean = checkBlogAdmin(userEmail); // 블로그 관리자
         const canEdit: boolean = isSameCommenter || !isLoggedIn || isBlogAdmin; // 수정 권한
-
+        const isVisibleConfirmDeletePassword =
+          deletingCommentId === commentId && checkingGuestPassword && !isLoggedIn; // 게스트 댓글 삭제
         return (
           <li key={commentId} className={`${styles.commentItem}`}>
             <div className={styles.thumbnail}>{isLoggedIn ? "✅" : "😀"}</div>
@@ -186,23 +208,18 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
                       onClick={
                         isLoggedIn || isBlogAdmin
                           ? () => handleClickDeleteButton(commentId)
-                          : () => handleClickGuestDeleteButton()
+                          : () => handleClickGuestDeleteButton(commentId)
                       }
                     >
                       삭제
                     </button>
                     <div
                       className={`${styles.guestConfirm} ${
-                        checkingGuestPassword && !isLoggedIn && styles.visible
+                        isVisibleConfirmDeletePassword && styles.visible
                       }`}
                     >
-                      <input
-                        type="text"
-                        placeholder="비밀번호"
-                        value={deletePassword}
-                        onChange={(e) => setDeletePassword(e.target.value)}
-                      />
-                      <button onClick={() => handleClickGuestPasswordConfirm(commentId)}>확인</button>
+                      <CustomInput placeholder="비밀번호" {...deletePasswordInputProps} />
+                      <button onClick={() => handleClickConfirmGuestPassword(commentId)}>확인</button>
                       <button onClick={() => handleClickCancelCheckingPassword()}>취소</button>
                     </div>
                   </div>
@@ -213,12 +230,7 @@ const CommentList = ({ postId, newUpdate, userEmail }: CommentListProps) => {
                 <p className={styles.date}>{getDateForm(date, true)}</p>
               </div>
               <div className={`${styles.editForm} ${editingCommentId === commentId && styles.editing}`}>
-                <textarea
-                  className={styles.textarea}
-                  placeholder="내용을 입력하세요"
-                  value={editComment}
-                  onChange={(e) => setEditComment(e.target.value)}
-                />
+                <CustomTextarea placeholder="댓글을 입력하세요." {...editCommentInputProps} />
                 <button onClick={() => handleClickCancelEditButton()}>취소</button>
                 <button onClick={() => handleClickConfirmEditButton(commentId)}>확인</button>
               </div>
